refactor(orchestrator): extract inline thread components for testability
Move ChatInput, MixedThread, ConnectionPill, LeanessIndicator, and
OrchestratorEmptyState into src/components/orchestrator/ as exported units.
